loop: Add bindings for `pw_loop`.
This adds a `Loop` struct for creating and managing a `pw_loop` struct.

It also adds a `LoopRef` struct that is a transparent wrapper around `pw_loop`,
and can be obtained from both a `Loop` and `MainLoop`.

The methods previously on the loop trait are moved there, and the loop trait is now only responsible for obtaining
such a `LoopRef` from any loop.